Krishi Ai/requirements.txt

Flask==2.3.3
google-generativeai==0.8.2
requests==2.31.0
